+++ liquibase-core/src/main/resources/META-INF/services/liquibase.command.CommandStep
@@ -50,4 +50,4 @@ liquibase.command.core.UpdateTestingRollbackCommandStep
 liquibase.command.core.UpdateToTagCommandStep
 liquibase.command.core.UpdateToTagSqlCommandStep
 liquibase.command.core.ValidateCommandStep
-liquibase.command.InternalDatabaseCommandStep
+liquibase.command.core.InternalDatabaseCommandStep
